+++ app/components/folio/console/ui/in_place_input_component.js
@@ -11,7 +11,7 @@ window.Folio.Stimulus.register('f-c-ui-in-place-input', class extends window.Sti
   }
 
   onBlur (e) {
-    if (e.detail && !e.detail.dirty) {
+    if (e.detail && !e.detail.dirty && !e.detail.failure) {
       if (this.hasAutocompleteValue) {
         const input = this.element.querySelector('.f-c-ui-ajax-input__input')